requirements.txt
alembic==1.7.1
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Excel==0.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
lml==0.1.0
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pyexcel==0.6.7
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
python-dotenv==0.17.0
pytz==2021.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.23
texttable==1.6.4
Werkzeug==2.0.1
WTForms==2.3.3
xlwt==1.3.0
zipp==3.5.0
numpy==1.18.4
FontAwesome==5.10.1.post1